refactor: PlayerBar 컴포넌트 구조 개선
- usePlayerProgress 커스텀 훅 분리
- formatTime 유틸리티 함수 분리
- 재생바 업데이트 주기 100ms로 변경하여 부드러운 움직임 구현
- 중복 상태(isSeeking, isDragging)를 interactionRef로 통합
