+++ app/(anon)/components/playerbar/utils/formatTime.ts
@@ -1,0 +1,8 @@
+// 시간을 분:초 형식으로 변환
+export const formatTime = (seconds: number): string => {
+  if (isNaN(seconds) || seconds === Infinity) return "0:00";
+
+  const minutes = Math.floor(seconds / 60);
+  const remainingSeconds = Math.floor(seconds % 60);
+  return `${minutes}:${remainingSeconds.toString().padStart(2, "0")}`;
+};
